+++ src/game/equipment/table.ts
