+++ src/property/_handler.ts
@@ -1,0 +1,46 @@
+import { Entity, World } from "@minecraft/server";
+import { IProperty } from "../declare/types";
+
+// World Property
+import PTeamScore from "./world/team_score";
+import PGameMode from "./world/game_mode";
+
+// Entity Property
+import PTeam from "./entity/team";
+import PAlive from "./entity/alive";
+import PMainWeapon from "./entity/main_weapon";
+import PSecondaryWeapon from "./entity/secondary_weapon";
+
+class WorldProperty {
+    private static _instance: WorldProperty;
+    static get instance() { return (this._instance || (this._instance = new this())); }
+
+    private propertyTable: Record<string, IProperty>;
+    constructor() {
+        this.propertyTable = {
+            "team_score": new PTeamScore,
+            "game_mode": new PGameMode
+        }
+    }
+    properties = () => this.propertyTable
+    get = (propertyId: keyof typeof this.propertyTable) => this.propertyTable[propertyId]
+}
+
+class EntityProperty {
+    private propertyTable: Record<string, IProperty>;
+    constructor(private entity: Entity) {
+        this.propertyTable = {
+            "team": new PTeam(this.entity),
+            "alive": new PAlive(this.entity),
+            "main_weapon": new PMainWeapon(this.entity),
+            "secondary_weapon": new PSecondaryWeapon(this.entity)
+        }
+    }
+    properties = () => this.propertyTable
+    get = (propertyId: keyof typeof this.propertyTable) => this.propertyTable[propertyId]
+}
+
+export default abstract class Property {
+    static world() { return WorldProperty.instance; }
+    static entity(entity: Entity) { return new EntityProperty(entity); }
+}
